requirements.txt
markdownify ~= 1.1
smolagents[openai]~=1.20
together ~= 1.5
python-dotenv ~= 1.1
spotify ~= 0.10
spotipy ~= 2.25.1
pytest-fastapi-deps ~= 0.2.3
langfuse ~= 3.2.1
openinference-instrumentation-smolagents~=0.1.14
opentelemetry-sdk~=1.35.0
opentelemetry-exporter-otlp-proto-http~=1.35.0
fastapi~=0.116
uvicorn[standard]~=0.35
python-multipart~=0.0.20